chore(communication): moving file to new exceptions folder
